better color organization
