+++ Assets/GameScripts/TileItems/ColorSchema.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: b1562698bdc3e5d4d8e497e35bb3fcc2
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
